[Flink] Make schema update support internal
