+++ flink/src/test/java/io/delta/flink/table/SchemaUpdateTestUtils.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (2026) The Delta Lake Project Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.delta.flink.table;
+
+import io.delta.kernel.types.StructType;
+
+/** Test-only access to package-private table primitives. */
+public final class SchemaUpdateTestUtils {
+
+  private SchemaUpdateTestUtils() {}
+
+  public static void updateSchema(AbstractKernelTable table, StructType targetSchema) {
+    table.updateSchema(targetSchema);
+  }
+}
